+++ app/src/main/java/com/example/googleclass/feature/post/data/model/TaskAnswerInPostModel.kt
@@ -5,10 +5,10 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class TaskAnswerInPostModel(
     val id: String,
-    val score: Int? = null,
+    val score: Float? = null,
     val submittedAt: String? = null,
     val status: String = "NEW",
     val files: List<AttachmentModel> = emptyList(),
-    val maxScore: Int? = null,
+    val maxScore: Float? = null,
     val postName: String? = null,
 )
